+++ spec/utils/index_spec.js
@@ -12,7 +12,7 @@ describe("Azk utils module", function() {
     });
     h.expect(current).to.not.equal(other);
     h.expect(current).to.equal(process.cwd());
-    h.expect(other).to.equal(__dirname);
+    h.expect(other).to.match(RegExp(h.escapeRegExp(__dirname), "i"));
   });
 
   it("should resolve a directory path", function() {
@@ -36,6 +36,57 @@ describe("Azk utils module", function() {
     var result, data = { value: "foo", hash: { key: "bar" } };
     result = utils.template("<%= value %> - #{hash.key}", data);
     h.expect(result).to.equal("foo - bar");
+  });
+
+  it("should return a array every call requireArray", function() {
+    h.expect(utils.requireArray()).to.eql([]);
+    h.expect(utils.requireArray(null)).to.eql([]);
+    h.expect(utils.requireArray(1)).to.eql([1]);
+    h.expect(utils.requireArray([1, 2])).to.eql([1, 2]);
+  });
+
+  describe("with function joinCmd", function() {
+    it("should return string passed", function() {
+      h.expect(utils.joinCmd("/bin/bash")).to.equal("/bin/bash");
+    });
+
+    it("should return a string that is the pices of the array", function() {
+      var cmd = ["/bin/sh", "-c", "ls"];
+      h.expect(utils.joinCmd(cmd)).to.equal("/bin/sh -c ls");
+    });
+
+    it("should escape args with spaces and quotes", function() {
+      var cmd, cmds = {
+        "/bin/sh -c \"echo 'hello world'\"": ["/bin/sh", "-c", "echo 'hello world'"],
+        "/bin/sh -c \"echo \\\"hello 'world\\\"\"": ["/bin/sh", "-c", "echo \"hello 'world\""],
+      };
+
+      for (var result in cmds) {
+        if (cmds.hasOwnProperty(result)) {
+          cmd = cmds[result];
+          h.expect(utils.joinCmd(cmd)).to.equal(result);
+        }
+      }
+    });
+  });
+
+  describe("with function splitCmd", function() {
+    it("should return a array with text pace", function() {
+      var str = "foo bar";
+      h.expect(utils.splitCmd(str)).to.eql(["foo", "bar"]);
+    });
+
+    it("should preserv single and double quotes", function() {
+      var cmd = [
+        `mongod --rest --text="hello world"`,
+        `--other='hello' "diferente other" -- -c "ls" echo "\\"hello\\""`
+      ].join(" ");
+      var expect_result = [
+        "mongod", "--rest", '--text="hello world"',
+        "--other='hello'", '"diferente other"', '--', '-c', '"ls"', 'echo', `"\\\"hello\\\""`,
+      ];
+      h.expect(utils.splitCmd(cmd)).to.eql(expect_result);
+    });
   });
 
   describe('envs', function () {