+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.16.2 - (2015-11-13)
+
+* Enhancements
+  * [Core] Upgrading node to v0.10.40
+  * [Suggesting] Changing suggestions to suggest a command based in Array in replace of String
+
+* Bug
+  * [Core] Fixing a bug that could generate an invalid command when `azk start` runs, making systems fail by not properly overriding the image Docker command by the custom command defined in the `Azkfile`;
+  * [Generators] Fixing replace evidences in `Court`;
+  * [Code] Fixing Docker-in-Docker entries in `Azkfile`;
+
 ## v0.16.1 - (2015-11-08)
 
 * Enhancements